fixed tests for new format
