+++ cdisc_rules_engine/check_operators/sql/matches_regex_operator.py
@@ -5,9 +5,14 @@ class MatchesRegexOperator(BaseSqlOperator):
     """Operator for regex pattern matching."""
 
     def execute_operator(self, other_value):
-        """target = self.replace_prefix(other_value.get("target"))
+        target_column = self.replace_prefix(other_value.get("target")).lower()
         comparator = other_value.get("comparator")
-        converted_strings = self.validation_df[target].map(lambda x: self._custom_str_conversion(x))
-        results = converted_strings.notna() & converted_strings.astype(str).str.match(comparator)
-        return results"""
-        raise NotImplementedError("matches_regex check_operator not implemented")
+
+        def sql():
+            return f"""CASE WHEN
+                            {target_column} IS NOT NULL 
+                            AND {target_column}::text ~ '{comparator}'
+                        THEN true
+                        ELSE false
+                        END"""
+        return self._do_check_operator(f"{target_column}_matches_regex", sql)
