+++ cdisc_rules_engine/check_operators/sql/not_matches_regex_operator.py
@@ -1,13 +1,18 @@
-from .matches_regex_operator import MatchesRegexOperator
+from .base_sql_operator import BaseSqlOperator
 
 
-class NotMatchesRegexOperator(MatchesRegexOperator):
+class NotMatchesRegexOperator(BaseSqlOperator):
     """Operator for inverted regex pattern matching."""
 
     def execute_operator(self, other_value):
-        """target = self.replace_prefix(other_value.get("target"))
+        target_column = self.replace_prefix(other_value.get("target")).lower()
         comparator = other_value.get("comparator")
-        converted_strings = self.validation_df[target].map(lambda x: self._custom_str_conversion(x))
-        results = converted_strings.notna() & ~converted_strings.astype(str).str.match(comparator)
-        return results"""
-        raise NotImplementedError("not_matches_regex check_operator not implemented")
+
+        def sql():
+            return f"""CASE WHEN
+                            {target_column} IS NOT NULL 
+                            AND NOT ({target_column}::text ~ '{comparator}')
+                        THEN true
+                        ELSE false
+                        END"""
+        return self._do_check_operator(f"{target_column}_not_matches_regex", sql)
